Add files via upload
Text mining and topic modeling are used to detect fraudulent behavior from the emails that Enron employees send. The project's goal is to find emails that have the topic of selling Enron stocks. Word search, sentiment analysis, word frequency finding, and topic analysis are done, respectively. The topic that mentions selling Enron stocks is revealed, and the emails that are related to that topic are found.